python/requirements.txt
# Viny2030 Python Requirements

# Backblaze B2
b2sdk==1.24.1

# Data Processing
pandas==2.0.3
openpyxl==3.1.2
xlrd==2.0.1

# API & Web
requests==2.31.0
python-dotenv==1.0.0

# GitHub API
PyGithub==2.1.1

# Utilities
python-dateutil==2.8.2
